fix(backend): move graphql execution in dedicated module
